refactor(app): update imports for layout/sections structure
