+++ constants/index.ts
@@ -1,13 +1,10 @@
-export * from './contact'
-export * from './navbar'
-export * from './hero'
-export * from './services'
-export * from './features'
-export * from './testimonials'
-export * from './pricing'
-export * from './cta'
-export * from './footer'
-export * from './gallery'
-export * from './faq'
-export * from './how-it-works'
-export * from './layanan'
+export * from "./contact";
+export * from "./layout/hero";
+export * from "./sections/features";
+export * from "./sections/testimonials";
+export * from "./sections/footer";
+export * from "./sections/faq";
+export * from "./sections/layanan";
+export * from "./sections/problem-awareness";
+export * from "./sections/whatsapp-float";
+export * from "./sections/testimonials-data-const";
